Release by Jason Duell =======================================================================
I've changed the default mappings in the script, to avoid 3-key maps.
Now it uses 'CTRL-' to do basic (no window split) cscope searches,
'CTRL-spacebar' to do searches that display results in a new,
horizontally split window, and 'CTRL-spacebar CTRL-spacebar' to do
searches with a vertical split.  I've also trimmed my overly verbose
comments in the file.

Since I can't update the description text on this site, I'm going to be
keeping the latest tutorial on vim-cscope usage at
http://cscope.sourceforge.net from now on.  I'll be posting a new
version of the tutorial there soon.

=======================================================================

Hmm, backslashes don't seem to show up when I type them on the upload
form.  What I meant to write in that last post was that the new default
mapping for regular cscope searches is 'CTRL-\'  (Control-backslash).'